requirements.txt
fastapi
pydantic
requests
sqlalchemy[asyncio]
asyncpg
pydantic-settings
alembic
crypto
passlib
bcrypt
python-jose